pyproject.toml
[tool.poetry]
name = "drinkit-stock-transfers"
version = "0.1.0"
description = "Sync Dodo transfers and push zero shipped rows to Google Sheets"
authors = ["Lev Korobeinikov <korobeinikoffl@yandex.ru>"]
packages = [{ include = "drinkit_stock_transfers" }]

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
flask = "^3.1.3"
requests = "^2.32.5"
python-dotenv = "^1.2.2"
cryptography = "^46.0.5"
psycopg2-binary = "^2.9.11"
gspread = "^6.2.1"
oauth2client = "^4.1.3"
aiogram = "^3.27.0"
aiogram-dialog = "^2.6.0"
apscheduler = "^3.11.2"
asyncio = "^4.0.0"
redis = "^7.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.6"
pytest = "^8.3.2"
pre-commit = "^3.7.1"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"